Prototype code for #c08-gloria
Arduino prototype code for #c08-gloria availing of UVC sterilization of taxi rear seats by way of detecting no human movement with PIR.
In practice, this would involve 'dual tech'. There should be a load sensor on the seat, in conjunction with the PIR. There should be no weight on the load sensors, nor movement within the field of view of the PIR in order for it to be deemed safe to trigger the UVC LED's to begin sterilization of the seating area.